some update on task 14'

+++ 0x03-user_authentication_service/app.py
@@ -33,34 +33,46 @@ def users():
     return jsonify({"email": email, "message": "user created"})
 
 
-@app.route('/sessions', methods=['POST'])
+@app.route('/sessions', methods=['POST', 'DELETE'])
 def login():
     """method to login user"""
-    email = request.form.get('email')
-    password = request.form.get('password')
+    if request.method == 'POST':
+        email = request.form.get('email')
+        password = request.form.get('password')
 
-    ok = AUTH.valid_login(email, password)
-    if not ok:
-        abort(401)
-    session_id = AUTH.create_session(email)
-    resp = make_response({"email": email, "message": "logged in"})
-    resp.set_cookie("session_id", session_id)
-    return resp
+        ok = AUTH.valid_login(email, password)
+        if not ok:
+            abort(401)
+        session_id = AUTH.create_session(email)
+        resp = make_response({"email": email, "message": "logged in"})
+        resp.set_cookie("session_id", session_id)
+        return resp
+    elif request.method == 'DELETE':
+        session_id = request.form.get('session_id')
+        if not session_id:
+            abort(403)
+
+        user = AUTH.get_user_from_session_id(session_id)
+        if not user:
+            abort(403)
+
+        AUTH.destroy_session(user.id)
+        return redirect('/')
 
 
-@app.route('/sessions', methods=['DELETE'])
-def logout():
-    """method to logout user"""
-    session_id = request.form.get('session_id')
+
+@app.route('/profile', methods=['GET'])
+def profile():
+    """method to register usersl"""
+    session_id = request.form.get("session_id")
     if not session_id:
+        
+        print('fula')
         abort(403)
-
     user = AUTH.get_user_from_session_id(session_id)
     if not user:
         abort(403)
-
-    AUTH.destroy_session(user.id)
-    return redirect('/')
+    return jsonify({"email": user.email}), 200
 
 
 if __name__ == "__main__":
